+++ CHANGELOG.md
@@ -11,6 +11,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - `FunctionError` type is replaced with `anyhow::Error`
 
+### Removed
+- Methods for adding rules and functions to a `RuleSet`, use the builder
+
 
 ## [0.4.4] - 2022-12-08